Temporary fix to correct ml2 bug.
Current packaging sets the ml2 configuration for the
neutron-plugin-openvswitch-agent, regardless of if you
want it. This temporary fix takes care of the issue. It
comments out the necessary lines, and does a manual stop/start
of the agent. Doing a service refresh of the agent does not
work, there needs to be a slight pause, accomplished by 'sleep 1'
in order for it to recognize the config change.